@@ -1,89 +1,72 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using BusinessService.Service;
-using Microsoft.EntityFrameworkCore;
-using Swashbuckle.AspNetCore.Swagger;
-using AutoMapper;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Serilog;
-using System.Collections.Generic;
+using Newtonsoft.Json;
+using CatalogApi.FilterException;
+using CatalogApi.Middlewares;
 
-namespace BusinessService
+namespace CatalogApi
 {
     public class Startup
     {
         public Startup(IConfiguration configuration)
         {
-            // Init Serilog configuration
-            Log.Logger = new LoggerConfiguration().ReadFrom.Configuration(configuration).CreateLogger();
             Configuration = configuration;
         }
 
         public IConfiguration Configuration { get; }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            var appSettingSection = Configuration.GetSection("AppSettings");
-            services.Configure<AppSettings>(appSettingSection);
-            
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
-            services.AddScoped(typeof(BusinessData.Repository.Interface.IBaseRepository<>), typeof(BusinessData.Repository.Concrete.BaseRepository<>));
-            services.AddScoped<IProductService, BusinessService.Service.ProductService>();
-            services.AddScoped<BusinessData.Repository.Interface.IProductRepository, BusinessData.Repository.Concrete.ProductRepository>();
-
-            services.AddScoped<IClientService, BusinessService.Service.ClientService>();
-            services.AddScoped<BusinessData.Repository.Interface.IClientRepository, BusinessData.Repository.Concrete.ClientRepository>();
-
-            var aa = Configuration.GetConnectionString("DefaultConnection");
-            
-            services.AddDbContext<BusinessData.BusinessContext>(options => options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection"), b => b.MigrationsAssembly("BusinessService")));
-            services.AddSwaggerGen(c => {
-                c.SwaggerDoc("v1", new Info { Title = "Business API", Version = "V1" });
-            });
-
-            var config = new MapperConfiguration(cfg =>
+            services.AddControllers(options =>
             {
-                cfg.CreateMap<BusinessData.Model.Client, Domain.ClientDTO>();
-                cfg.CreateMap<Domain.ClientDTO, BusinessData.Model.Client>();
-                cfg.CreateMap<IEnumerable<BusinessData.Model.Client>, IEnumerable<Domain.ClientDTO>>(MemberList.Destination);
-                cfg.CreateMap<BusinessData.Model.Product, Domain.ProductDTO>();
-                cfg.CreateMap<Domain.ProductDTO, BusinessData.Model.Product>();
-                cfg.CreateMap<IEnumerable<BusinessData.Model.Product>, IEnumerable<Domain.ProductDTO>>(MemberList.Destination);
-                cfg.CreateMap<BusinessData.Model.Category, Domain.CategoryDTO>();
-                cfg.CreateMap<Domain.CategoryDTO, BusinessData.Model.Category>();
+                options.Filters.Add<GlobalExceptionFilter>();
             });
-
-            var mapper = config.CreateMapper();
-            services.AddSingleton(mapper);
+                /*
+            AddNewtonsoftJson(option =>
+            {
+                option.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+                option.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
+            });
+                */
+            services.AddCors();
+            services.AddServicesApp();
+            services.AddSwagger();
+            services.AddDbContexts(Configuration);
+            services.AddMapper();
+            services.AddValidator();
 
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, Microsoft.Extensions.Logging.ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
             if (env.IsDevelopment())
             {
-                app.UseDeveloperExceptionPage();
+                app.UseMiddleware<ExceptionMiddleware>();
+                app.UseSwagger();
+                app.UseSwaggerUI(c => {
+                    c.SwaggerEndpoint("/swagger/v1/swagger.json", "post API V1");
+                });
             }
-            else
-            {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-
+            
             app.UseHttpsRedirection();
-            app.UseMvc();
-            app.UseSwagger();
-            app.UseSwaggerUI(c => {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "post API V1");
+            app.UseRouting();
+            /*
+            app.UseCors(option =>
+            {
+                option.WithOrigins("http://localhost:8080/");
+                option.AllowAnyMethod();
+                option.AllowAnyHeader();
             });
-            loggerFactory.AddSerilog();
-
+            */
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
