@@ -4,13 +4,14 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace BusinessData.Model
+namespace CatalogData.Model
 {
     public class Product : BaseEntity
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int IdProduct { get; set; }
+        public string ShortDescription { get; set; }
         public string Description { get; set; }
         public decimal UnitPrice { get; set; }
         public int UnitStock { get; set; }
@@ -19,7 +20,7 @@ namespace BusinessData.Model
 
         public virtual Category Category { get; set; }
 
-        public virtual IEnumerable<InvoiceDetail> InvoiceDetail { get; set; }
+        
 
     }
 }
