@@ -6,8 +6,9 @@ using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
-namespace BusinessService.Controllers
+namespace CatalogApi.Controllers
 {
+    [ApiController]
     [Route("api/[controller]")]
     public class CategoryController : ControllerBase
     {
