@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace BusinessService.Domain
+namespace CatalogApi.Domain
 {
     public class CategoryDTO
     {
