@@ -7,54 +7,75 @@ using System.Linq;
 using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
-using UserService.Domain;
+using UserApi.Domain;
+using UserApi.Helper;
+using UserData.Repository.Interface;
+using System.Linq.Expressions;
+using UserData.Models;
 
-namespace UserService.Service
+namespace UserApi.Service
 {
     public class AuthenticateService : IAuthenticateService
     {
         private readonly AppSettings _appSettings;
-        private List<Login> _login = new List<Login>
-        {
-            new Login { id = 1, firstname = "Ajaya", lastname = "Nahuel", username = "CNA", password = "123" },
-            new Login { id = 2, firstname = "AAAAA", lastname = "Aaaaa", username = "AAA", password = "123" }
-        };
-
-        public AuthenticateService(IOptions<AppSettings> appSettings)
+        private readonly IGenericRepository<User> _genericRepositoryUser;
+        private readonly IGenericRepository<TokenUser> _genericRepositoryToken;
+        public AuthenticateService(IOptions<AppSettings> appSettings, IGenericRepository<User> genericRepositoryUser, IGenericRepository<TokenUser> genericRepositoryToken)
         {
             this._appSettings = appSettings.Value;
+            _genericRepositoryUser = genericRepositoryUser;
+            _genericRepositoryToken = genericRepositoryToken;
         }
 
-        public Domain.SecurityToken Authenticate(string username, string password)
+        public async Task<TokenUser> Authenticate(string username, string password)
         {
-            var user = _login.SingleOrDefault(x => x.username == username && x.password == password);
+            
+            string passwordEncrypt = Encrypt.GetSHA256(password);
 
-            if (user == null)
-                return null;
+            Expression<Func<User, bool>> funcPred = p => p.Username == username && p.Password == passwordEncrypt;
+
+            var user = await _genericRepositoryUser.FirstOrDefault(funcPred);
+
+            if (user == null) return null;
+
             // authentication successful, generate jwt token
+            return await GenerateToken(user);
+            
+        }
+
+        private async Task<TokenUser> GenerateToken(User model) 
+        {   
             var tokenHandler = new JwtSecurityTokenHandler();
             var key = Encoding.ASCII.GetBytes(_appSettings.Secret);
             var tokenDescriptor = new SecurityTokenDescriptor
             {
                 Subject = new ClaimsIdentity(new Claim[]
                 {
-                    new Claim("username", username),
-                    new Claim("firstname", user.firstname),
-                    new Claim("lastname", user.lastname)
+                    new Claim(ClaimTypes.NameIdentifier, model.Username),
+                    new Claim(ClaimTypes.Name, string.Format("{0}, {1}",model.FirstName, model.LastName))
                 }),
-                Expires = DateTime.UtcNow.AddMinutes(1), // Se configura para que expire dentro de 1 minuto, sacar luego
+                Expires = DateTime.UtcNow.AddHours(3), // Se configura para que expire dentro de 3 horas, sacar luego
                 SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
             };
+
             var token = tokenHandler.CreateToken(tokenDescriptor);
             var jwtSecurityToken = tokenHandler.WriteToken(token);
 
-            return new Domain.SecurityToken() { auth_token = jwtSecurityToken,expire = tokenDescriptor.Expires };
+            TokenUser tokenUser;
+            //Registro en base el token para el usuario con su expiracion
+            
+            tokenUser = new TokenUser()
+            {
+                IdUser = model.IdUser,
+                Token = jwtSecurityToken,
+                InitDate = DateTime.Now,
+                EndDate = tokenDescriptor.Expires
+            };
+            await _genericRepositoryToken.AddAsync(tokenUser);
+
+            return tokenUser;
         }
 
-        public IEnumerable<Login> GetAll()
-        {
-            return _login;
 
-        }
     }
 }
