@@ -2,15 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using UserService.Domain;
+using UserApi.Domain;
+using UserData.Models;
 
-namespace UserService.Service
+namespace UserApi.Service
 {
     public interface IAuthenticateService
     {
-        SecurityToken Authenticate(string username, string password);
-
-        IEnumerable<Login> GetAll();
-
+        Task<TokenUser> Authenticate(string username, string password);
     }
 }
